Update INSTALL.md list of (un)supported platforms
This commit adds a list of NON supported platforms to `INSTALL.md`.

A "soft" disclaimer of support for macOS/aarch64 is added to the new
list, resolving issue 411.

A blanket disclaimer for Clang and GCC forks from several vendors is
added.  This includes Intel's `icx` and `icpx`, and so this commit
resolves issue 458.  The pre-existing disclaimer of support for Arm
Ltd.'s forks of Clang and GCC has been absorbed into this item.

A full disclaimer of support for NVIDIA-branded compilers is added.
Unlike the case with Arm Ltd.'s compilers, this commit preserves the
duplicate texts under the PGI compilers bullets for Linux/x86_64 and
Linux/ppc64le.