Resolve issue 430: cannot disable the default network
This commit adds logic to parse the configure arguments (inclusive of
the `GASNET_CONFIURE_ARGS` inherited from the evironment) for options
which would disable the default network.  If such is found, an error
message gives two options: pick a different default or remove the
disable option.

Care is taken to allow `--disable-smp -enable-smp` by examining only
the last of any sequence of the enable and disable options (with
single- or double-dash).