fixup: Resolve issue 430: cannot disable the de...
This commit corrects a behavioral regression introduced while fixing
issue 430.  The new problem was that a configure option such as
`--disable-ibv-odp` was matching `(dis|en)able-$GASNET_CONDUIT` in
a regex.

The solution is to observe that arguments in `$GASNET_CONFIGURE_ARGS`
are followed either by a space (if they came from the inherited
environment) or by a double-quote (if they came from the command line
or were otherwise injected by the configure script).  The potential
case of ending immediately following an argument from the inherited
environment cannot occur because we always append at least two
options: `--with-{c,cxx}flags="..."`.  However, a trailing space has
been appended to the `egrep` input to avoid depending on that.

Additionally, the case of `-enable-ibv=[FOO]` is considered so that
`-enable-ibv=no` counts as a disable (and any other `[FOO]` enables).