test/copy: issue #241: Add missing CUDA device synchronization
The test initializes device memory using `cuMemcpyHtoD()`, but fails to
synchronize the transfer from pageable host memory to device using
the default stream. Although this call is "synchronous", as described in:
  https://docs.nvidia.com/cuda/cuda-driver-api/api-sync-behavior.html
this means the device memory might not be fully updated before the
copy call returns. That device memory is subsequently accessed by `copy()`
which accesses the device memory directly (via GDR) and/or via
reference kinds using `cuMemcpyDtoHAsync()` on an independent
(unordered) stream.

This missing synchronization means there was a race on device memory
that *might* explain the intermittent failures reported in issue #241.

Fixes issue #241 (I hope).