Implement and document `make install NETWORK=...`
This commit makes a relatively small change to the installation logic
to allow `make install NETWORK=...` to specify a default network
(where previously setting `UPCXX_DEFAULT_NETWORK` was required).
Additionally, this capability is now documented in `INSTALL.txt` and
in the output at the end of `all`, `check` and `install` steps.