copy: Peel 3rd party copy case into a separate function
This moves the code for the special case of a third-party copy (where
the initiator is neither the source or destination rank) into a
separate no-inline function.

This case needs to be correct, but is hopefully/presumably an uncommon
case, because it pays unavoidable extra communication costs to
orchestrate the transfer. Moving the code to handle it into a separate
no-inline function hopefully reduces icache pressure on the common,
more performance-critical cases.

The diff for this commit looks "messy", but it really just moves
the 3rd party code to a new function and hoists the special-case
check above redundant detail::copy preamble.