copy: Add private promotion optimization
For copy overloads accepting a `T*` (pointer-to-local) we previously
assumed this always pointed to private memory. When the other pointer
references device memory, three of the four point-to-point cases require
an extra bounce-buffer copy due to the private heap.

Now cases involving device memory and a raw `T*` will attempt to
upcast/promote the `T*` (as with `try_global_ptr(T*)`), in case it
happens to point into a local host segment. When successful, this
can allow us to elide the extra bounce-buffer copy.

This optimization might eventually be retired, once GASNet memory kinds
is uniformly available on all conduits and bug 4148 is resolved.