This comment was incorrect about memcpy and TriviallyCopyable types. As of C++ standards through the present, memcpy can never create an object.
The bug of issue #400 being worked around was also being caused by a combination of undefined behaviors: object creation and pointer aliasing rules. The comment only considered the former.

This comment was updated to correctly identify undefined behavior, the C++17 standard tools necessary to fix the second type of UB, and the paper currently in committee which will allow the first type of UB to be fixed. This will better allow us to track the status of necessary C++ standard support so that a compliant implementation can be enabled when compiling with a sufficient C++ standard.