Move default network selection to post-configure
This commit resolves issue 354 (improve default network selection)
by deferring selection of a default network conduit in most cases.

Now only cross-compilation for Cray systems, and the use of
`--with-default-network=...` establish a default network at configure
time.  For all other cases, the default network is established
after GASNet is configured, by taking the "best" network which is
officially supported by UPC++ and located by GASNet's configure.

The original plan was to use a script in `utils/config/gasnet`, but
that turns out to have a circular dependency: those scripts are to be
run with the environment mirroring the settings in the makefile
fragment of the (not yet determined) default network.  So, the
approach taken in this commit is to inline the selection logic into
`bld/Makefile` to run *between* GASNet's `configure` and the runs of
the scripts in `utils/config/gasnet`.