Add network-specific sanity checks
This commit adds `conduit_warnings.sh` to the set of scripts run
immediately following GASNet's `configure` to generate values in the
corresponding `gasnet.{codemode}.mak`.  The purpose of this script is
to populate `$(UPCXX_CONDUIT_WARNINGS)` with text to be output at the
end of `all`, `check` and `install` steps along with the listing of
default and available conduits.

This initial version parses the `ibv_devinfo` output looking for
active InfiniBand ports, and populates `$(UPCXX_CONDUIT_WARNINGS)`
with an appropriate message if `ibv_devinfo` could be run, but did not
report any such ports.  Additionally, OPA (Omni-Path) HCAs are
identified and excluded from the search for active ports.  However, if
only OPA HCAs are located then a *different* warning is issued to
indicate that mpi-conduit is probably a more appropriate choice.