Add work-arounds for "nvhpc" compilers.
This commit follows-up on GASNet-EX work (GASNet PR 404) by adding
recognition of "nvhpc" as a compiler family in `compiler.mak`
AND applying all the same customizations as for the PGI family.

This work does *not* constitute "adding support" for this compiler
family.  It only provides the minimum requirements to perform the
evaluation required to make such a change.