configure: add GASNET_CONDUIT as a reported envvar
This commit adds `GASNET_CONDUIT` to the list of environment variables
for which `configure` reports the value it sees.  While we
intentionally lack corresponding user-facing documentation, the value
of this variable at configure time determines the default network in
the absence of a `--with-default-network=...` option.  So, reporting
its value at configure-time may be necessary to properly diagnose an
end-user's problem report.