Improve conduit-dependent static locality information
Hoist `backend::all_ranks_definitely_local` from a `const` variable
defined in libupcxx.a to a `constexpr` variable visible in the headers.

In particular, this should provide normal (non-LTO) optimization
sufficient information to statically optimize the expression in
`backend::rank_is_local` to remove what is currently a dynamic branch.
In particular, this should allow the optimizer to compile away the
`backend::rank_is_local` locality branches in rput/rget added in
PR 345 when building for smp-conduit, and discard the code for the
remote branch that is now statically known to be unreachable on this
conduit.

This optimization became possible with the introduction of
upcxx_config.hpp, which came after this was originally written.