Fix issue 479: intermittent lpc-stress/opt failures on ARM64
Turns out this was not actually a hang, but rather a pathological
slow-down caused by thrashing the memory system.

After reproducing the slow-down manually on a wombat login node (Cavium
ThunderX2), I used gdb to confirm the master thread was spending
most of its time trying to enqueue LPC acknowledgements to producer
threads who are spinning on burst (dequeue).

Inserting padding to separate the head and tail pointers in the
atomic MPSC queues seems to reliably solve the problem.

Even on architectures where this wasn't causing a critical slow-down,
it's likely this defect was degrading the performance of our
multi-threaded LPC queues any time they exceeded one element. The
queue algorithm was designed to allow the head and tail to proceed
independently, but with both pointers on the same cache line false
sharing between the producers and consumer was likely leading to lots of
unnecessary coherence cache misses.