Implement and document nvhpc compiler support
This commit updates documentation and install-time validation logic to
support the NVIDIA HPC SDK (aka "nvhpc") compilers on x86_64 and
ppc64le platforms.  Support for aarch64 might be added in the future.

In documentation, the mentions of PGI now state "through 20.4" and the
text "or any later versions" in configure-time output is adjusted by
appending "where no end-of-range is given".  Meanwhile, documentation
of support for nvhpc lists a start version of 20.9.  Together these
implicitly exclude the very broken 20.7 release without the need to
list it specifically.

In the `platform_sanity_checks()` bash function, the logic has been
updated to treat PGI and nvhpc as a single family which has an
excluded version range covering 20.5 through 20.8 (though only 20.7 is
know to exist).