Improve configure processing of CC, CXX and flags
This commit resolves issue 496 by making two groups of changes to the
handling of `CC`, `CXX`, `CFLAGS` and `CXXFLAGS`.

The first set of changes is to add missing `eval` to several compiler
invocations in `utils/system-checks.sh`.  These are necessary to deal
properly with quotation marks appearing in these variables' values,
especially when used to preserve white space.

The second change is to modify the mechanism by which the compiler
settings are passed to GASNet's `configure` to match that already
in use for the flags.

The result is that the following contrived example command runs to
completion and the subsequent `make` passes the correctly quoted
arguments to GASNet's `configure` (which will choke on white space in
`CC` and/or `CFLAGS`, but that is an autoconf problem, not in-scope
for support in UPC++ or GASNet).

```
$ GASNET_CONFIGURE_ARGS='--with-foo="a horse walks into a bar"' \
  ../configure \
        --without-mpicc \
        CXX='g++ -Dpart1="Twinkle, twinkle,"' \
        CFLAGS='-Dpart2="little star."' \
        --with-cc='gcc -Dpart3="How I wonder"' \
        --with-cxxflags='-Dpart4="what you are."'
```