configure probe logic for Intel compilers on macOS
This commit generalizes configure-time checks of the Intel Classic
compilers to allow for both gcc-based (Linux) and clang-based (macOS)
backend toolchains.

Prior to this, attempts to use the Intel compilers on macOS would fail
the gcc toolchain probe.  That is no longer case, but rather than
allowing the Linux-centric probe to accept this compiler, new
macOS-specific logic ensures that (for now) this target is categorized
as "has not been validated", rather than as good or bad.

The new logic places the floor at 2021.3.  Though older versions are
known to exists, our testing has been exclusively on 2021.3.0.