Ensure correct cleanup of operation_cx objects when `rpc` throws
Previously throwing from rpc would result in dropping the operation_cx
completions on the floor. This led to memory leaks of the completion
objects, skipped destructors for user-owned objects passed to
completions, and missing decrements on cancelled `as_promise`
completions.

This required deploying a new cancellation "virtual" function
for `lpc_dormant` objects, which is populated when contructing
the completions and invoked during unwinding when an exception
is thrown from injection.

Currently cancellation handling is only performed for `rpc` operation_cx
completions. It may also need to be applied elsewhere.