Add Cross-Code Segment (CCS) RPC support
This commit is documented at docs/ccs-rpc.md

This patch adds support for relocating function pointers to other
processes across multiple dynamic shared objects (DSOs). Relocation off
`global_fnptr_basis` is replacecd with relocation off of the start of
the segment containing the function pointer. This allows for the safe
relocation of function pointers in segments other than the "primary"
UPC++ segment, formerly the segment containing `global_fnptr_basis`, in
the presence of address space layout randomization (ASLR).

The relocation mechanism is designed to create various classes of
"function tokens," which contain the information necessary to identify
segments in a uniform manner across processes and an offset to retrieve
the relocated function pointer. There are three such "function token"
classes: `function_token_ss`, `function_token_ms`, and `function_token`.
These are designed to plug into `global_fnptr` as template parameters to
choose between supported relocation types.  For instance, a
`global_fnptr<F,function_token_ss>` will use a relocation strategy
equivalent to that used before this patch.

Function Tokens

`function_token_ss` preforms a "single segment" relocation against a
single universally agreed upon "primary" segment. This causes relocation
to occur in an equivalent manner to the strategy this patch replaces,
using just a single `uintptr_t` offset to communicate function pointers
for relocation. By default, this uses the address of the start of the
segment rather than the address of a basis function. The primary UPC++
segment is identified as the segment containing a basis function.

`function_token_ms` preforms a "multi segment" relocation against an
executable segment of a DSO. This token type sends a hash to identify
the containing segment and an offset to retrieve the pointer into that
segment. `segmap_cache` is responsible for mapping and lookups of
dynamic segments. By default, `function_token_ms` uses the
`segmap_cache` instance stored in `persona_tls` for its first level
cache.

`function_token` provides a smart function relocation token which
becomes a `function_token_ss` if the provided pointer is within the
range of the primary segment. Otherwise, it becomes a
`function_token_ms`.  The type of function token stored can be queried
with `function_token::token_ident()` for more efficient serialization.

`function_token_ss` is the default function token type unless a
translation unit is compiled with `-DUPCXX_CCS_RPC=1`.

`segmap_cache`

The meat of the segment mapping and caching takes place in the
`segmap_cache` class. The main components of this class are the level 1
cache, level 2 cache, and detailed segment map used for debugging and
building the caches.

The level 1 cache is intended to be used at thread level and is not
thread safe to avoid the overhead of locking and atomic operations. An
instance of `segmap_cache` is created in the `persona_tls` for this
purpose.  Due to the restrictions on constructors and destructors of
thread local storage, the level 1 cache uses `std::array`s for storage.
Two sorted arrays are used to enable binary searches of the level 1
cache in each direction. The capacity of this storage can be set with
`UPCXXI_MAX_LEVEL1_SEGCACHE_SIZE`, which defaults to 20.  This cache
only contains segments that the program has actually used, so in order
to exceed this threshold, the program would have to not only use at
least 20 libraries, but also make direct RPC calls to function pointers
within them. In practice, even if that many libraries were loaded, most
would likely not have function pointers invoked directly.

The level 2 cache is held as a process-wide static structure and is only
used if the level 1 cache is full.  It contains all mapped segments. If
there is a miss at level 1 cache and the level 1 cache is not full,
lookup goes directly to the heavy-weight segment map that can be used to
add the segment to the level 1 cache. While it would be possible to
promote a level 2 segment to level 2, rebuilding the level 2 cache is
not thread safe and requires the user to call `rebuild_cache()`. The
segment map, on the other hand, uses locks and can be rebuilt
automatically if a library was loaded. Because hitting the uncached
segment map would only happen once per thread per segment, there is
negligible benefit to maintaining an additional cache promotion
mechanism.

The segment map contains information necessary to build the caches as
well as information used for debugging, such as flags to indicate the
status of a segment and pointers to the Elf symbol and string tables for
Elf binaries. It is built by `segmap_cache::build_segment_map()`, which
uses platform-specific code to parse the process's program headers. This
is currently implemented for Elf and Mach-O file formats. These
mechanisms still work if the program is statically linked, simply
resulting in the expected two segments of the main program and the vDSO
injected by the kernel. This platform-specific behavior can be disabled
by configuring UPC++ with `--disable-ccs-rpc`, which causes the map to
contain a single "segment" of the range of the address of the basis
function to the address of the basis function plus one.  This allows it
to be picked up as the primary segment.

`segmap_cache` also contains several static member functions for
debugging tokenization. Output is colorized if `isatty` detects a
terminal, but can be overridden with the `UPCXX_COLORIZE_DEBUG`
environment variable.  The dynamic linkage debug functions print tables
of the mapped segments, the attempted lookup, and the symbol of the
function if available. An executable must be built with `-rdynamic` for
its symbols to be exported for this mechanism to find.

It is assumed that libraries are not unloaded, or at least not unloaded
and the address space reused by another library.  The level 1 caches are
not purged if a library is unloaded as `rebuild_cache()` only affects
the level 2 cache and segment map and is unable to touch every thread's
thread-local storage. In practice, `dlclose` usually doesn't actually
unmap the library and this shouldn't be a limitation with any practical
effect.

While developing this PR, some usage of `FunctionToken`s were
erroneously defaulting to `FunctionTokenType` within libupcxx, which is
supposed to be controlled by the user with `UPCXX_CCS_RPC=1`.  To
prevent bugs, `FunctionTokenType` is set to the incomplete class
`function_token_invalid`, which forces compile errors if the defaulting
of these template parameters would propagate into libupcxx.