Add branch prediction hints to UPCXX_ASSERT(_ALWAYS)
Ensure the optimizer correctly branch-predicts assertions to succeed,
especially for the case of `UPCXX_ASSERT_ALWAYS` which incurs a
dynamic branch in production mode.

Some optimizers were probably already inferring the right behavior
here due to the fact the assertion failure path invokes a function
annotated with `__attribute__((__noreturn__))`, but this makes the
optimizer hint explicit.