Issue 490: Move destruction into function and inline destructor
Attempting to fix the pedantic warning about C++ destructor name to
comply strictly with ISO C++ results in hard compiler errors with
EDG-based compilers. Moving the "meat" of the destructo into a separate
function to preserve "coldness" of call and inlining the actual
destructor fixes this out-of-line destructor definition warning in a
way that shouldn't error on broken compilers.