Issue #512: Allow ADL with when_all
`when_all` was failing to be found as expected with ADL. This is because
upcxx::future is just an alias to upcxx::detail::future1 and ADL does
not examine the parent namespace. The way UPC++ is documented, one would
expect ADL to work:

auto fut = upcxx::make_future(42);
when_all(fut);

This is how `when_all` is expected to work with the upcoming Executor
proposal: `std::when_all` would be written unqualified (just `when_all`)
to invoke ADL and find cusomized implementations. The same mechanism can
be used with `begin(vector)` (which will find `std::begin` for STL
vectors but will also function without modification if the replaced with a
3rd party vector implementation. `std::get` is another where ADL can be
used to make code more generic.

fixed when_all_fast