upcxx-run: add a network option
This commit implements a `-network` option for `upcxx-run` which
enables it to bypass the normal step of extracting the network from
the executable.

Other than the usage output, this option is intentionally undocumented
because at this point in time we believe the only usage cases are ones
we do not yet support (such as dynamic loading of a shared libupcxx).