Silence LTO-specific warnings
When LTO and `-Wall` are used together with (at least) GCC 11.2.0, the
compiler is able to observe that `gex_Segment_QueryBound()` will not
write to the three output locations when there is no bound segment.
This results in "may be used uninitialized" warnings at application
link time.

This commit silences these warnings by providing zero-initialization
for the variables in question.

Note that this is startup code outside performance-critical paths.