detail::rma_copy_local(): Move CUDA-specific parts to cuda.cpp
The CUDA-specific block of code in copy.cpp:detail::rma_copy_local() is moved
to cuda.cpp:detail::cuda_copy_local(). This is part of an overall effort to
partition device-specific code in separate source files, to defend
against the possibility that some device libraries might provide headers
that conflict with those of other devices, or require conflicting compiler flags.

There are no logic changes in this commit, just simple code movement.

The additional level of function call introduced by this motion will be
removed in a subsequent commit.