Add shell-based compilation of tests
If a `*.sh` file is found in a test directory, then it is run as a
bash shell script to build a UPC++ executable.  It runs with the
`UPCXX_{CODEMODE,THREADMODE,NETWORK}` environment variables set (among
others), and the correspondingly named target executable as the only
argument.

Maintainer-facing docs have been added to `docs/maintainer.md`, which
is supported by a new `test/hello_via_shell.sh` example.

The implementation includes a TODO note regarding the possible
addition of support for `.py`, `.pl` and `.mak` suffixes.  The
implementation is designed such that adding the conditional
interpreter setting is likely to be less work than will be required to
strip additional suffixes via make.  Since the UPC++ and GASNet-EX
configure scripts locate  python, perl and gmake, these additions are
not unrealistic if the demand arises.