configure: error handling in CC/CXX mismatch probe
This commit improves the logic which probes for mismatch between the
family or version of `$CC` and $CXX`.  A return code other than 0 or 1
from the respective probes indicates a failure mode other than
mismatch, such as failure to run the preprocessor.  The error handling
has been updated to report this as distinct from the mismatch case.