configure: resolve issue 521
This commit resolves issue 521: "compiler mismatch logic broken in CI
(regression)" by correcting the handling the `--enable-single=...`
case in which `$GASNET` references a build directory.

The logic introduced as `cpp_extract_pp_expr()` by pull request #396
erroneously assumed the two possible cases were source or install.
However, the second case is a *build* directory, from which one needs
to query the top-level GASNet-EX source directory.  This is query done
by extracting the value of `TOP_SRCDIR` from the top-level `Makefile`.

The following alternative was considered, but rejected because it
would require changing `configure` to swap the order in which one
sets up and validates `$GMAKE` versus the compilers:

```
  eval "$( $GMAKE -f $GASNET/Makefile echovar VARNAME=TOP_SRCDIR )"
```