{cuda,hip}.hpp: Add a comment to record a resolution
As resolved in Slack on 2022-02-11, we only increment these feature
macros as part of the release procedure when we are issuing a release
with a corresponding spec revision that alters their specified value.
In particular, they are NOT incremented in intervening development
snapshots.

Codes seeking finer-grained insight into when features were deployed
during development between releases should consult `UPCXX_VERSION`
bumps in git history.