test/memberof: Deploy new GPU type aliases
This program is a correctness test and not intended as an example code.

However the diff for this commit is a good illustration of how the
new `gpu_default_device`/`gpu_heap_allocator` type aliases and
`make_gpu_allocator` can simplify a non-contrived code that is written
to function with either GPU kind.

The diff notably omits the 10 lines of factored preprocessor goop in
test/util.hpp that defines the DEVICE macro used to support "the bad old way".