Assert correct return from custom deserialize()
The UPC++ spec mandates:

`deserialize()` must use placement new to construct the resulting object in
the provided storage and return a pointer to the object.

The runtime was not enforcing this requirement, but WAS silently
relying upon it in several places.

Add assertions to detect violations of this invariant in user-provided
custom deserialization functions.