Kinds: Retire device_allocator_core tombstones
These tombstone objects are no longer needed, as we now enforce a
precondition that device_allocator destruction requires device destroy.
This means it's no longer possible to destruct a device_allocator and
leave an active device that could be passed to the constructor of a new
device_allocator, which is what these tombstones protected against.