Correct test/ccs/Makefile for GNU Make 3.80
This commit removes the use of the `abspath` and `lastword` functions
from `test/ccs/Makefile`.  Both were introduced in GNU Make 3.81,
while our floor version is currently 3.80.

In the case of `abspath`, the function has simply been removed.  This
means that instances of `..` and `.` (if any) might be preserved in
the value of some directory variables.  While this might be less
pleasing cosmetically, it does not change the semantics in any way.

In the case of `lastword`, the following replacement has been made:
  `$(lastword $(FOO))`  => `$(word $(words $(FOO)), $(FOO))`