CCS: Compress verified multisegment relocations
Because verified segments can be arranged in a deterministic order,
all ranks have the same list of verified segments and by sorting each
batch of verifications, segments can be assigned incremental IDs.
Instead of using a hash, where enough bits must be used to make
hash collissions extremely unlikely, this unique ID can fit within just
a few bits. A vector of these sorted verified segments is stored such
that their index in this vector matches their ID, leading to fast lookup
on the receiver. To enable this new compressed method of multisegment
relocation, a new function token type `function_token_ms_idx` is
introduced. This type can be serialized using just 64 bits, the same as
a single segment relocation.  The bottom 48 bits are used to store the
offset, more than enough to cover the size of a `.text` segment. The
next 15 bits are used to store an `int16_t` worth of segment indexes,
again far more than any resonable program would have. The MSB of the
encoding is set to 1, indicating a multisegment relocation.

The special case of an ID of zero indicates an unverified segment. This
causes a `function_token_ms` (which uses an offset and hash) to be used
by `function_token` instead.

In terms of performance, there is a tradeoff of using a verification
reduction vs. reducing message size.