Fix incorrect treatment of two tests
This commit resolves issue #542 in which is was observed that `make
tests` and `make check` would compile `test-rput_thread` and
`test-uts_hybrid`, but `make check` did not run them and `make
tests-clean` did not remove them.

The problem was an inadvertent reversal of seq-vs-par in the logic to
generate filtered lists of tests.