Avoid a dangling-pointer warning from GCC 12+
In testing the brand-new GCC 12.1.0 release, I have encountered a new
warning building UPC++ with `-Wall`.

The new GCC adds a `-Wdangling-pointer[=n]` warning to the frontend
for the C family of languages.  Additionally, `-Wall` now enables
`-Wdangling-pointer=2`.  In my testing this results in exactly one
new warning when compiling UPC++ with `-Wall`:

```
/[redacted]/upcxx/src/future/core.cpp: In member function 'void upcxx::detail::future_header::entered_ready_with_sucs(upcxx::detail::future_header*, dependency_link*)':
/[redacted]/upcxx/src/future/core.cpp:103:18: warning: storing the address of local variable 'active_head' in '{anonymous}::active_tail_' [-Wdangling-pointer=]
  103 |     active_tail_ = &active_head;
      |     ~~~~~~~~~~~~~^~~~~~~~~~~~~~
/[redacted]/upcxx/src/future/core.cpp:100:28: note: 'active_head' declared here
  100 |   future_header_dependent *active_head = nullptr;
      |                            ^~~~~~~~~~~
/[redacted]/upcxx/src/future/core.cpp:41:38: note: '{anonymous}::active_tail_' declared here
   41 |   __thread future_header_dependent **active_tail_ = nullptr;
      |                                      ^~~~~~~~~~~~
```

As a human I can see that the value of `active_tail_` is never equal
to `&active_head` at the end of `entered_ready_with_sucs()`, even if
the compiler cannot.  So, this commit adds a `UPCXXI_ASSUME()` to
inform the compiler of this fact, which is sufficient to eliminate the
warning.