issue 550: ROCm/HIP headers break the GNU `__noinline__` attribute
Workaround the ROCm/HIP header which breaks the GNU `__noinline__`
function attribute in a way that leads to compile errors with `hipcc`.

The workaround is to detect the presence of the header containing the
objectionable macro definition and disable our own use of the attribute
which the header has corrupted.

Resolves issue #550