tests: fix use of timeout utility
This commit modifies the invocation of the GNU coreutils `timeout`
utility used to limit running time of hung (or unacceptably slow)
tests runs.

Previously the `-k` option was passed an argument which was the sum
of the base time limit and a "pad".  The intention of the pad is allow
for normal exit (in response to `SIGTERM`) when such works, but to use
`SIGKILL` if it does not.  However, the documentation (since at least
coreutils-8.22) documents the time for the kill behavior as starting
*after* the original timeout, making the sum erroneous.  Now, we just
pass the pad as the argument to `-k`.