INSTALL.md: update known issues with CUDA memory kinds
The `--enable-issue557-workaround` configure option exists to work
around two GASNet-EX bugs: 4485 and 4494.  The former is not present
in any current GASNet-EX, and the latter has an effective work around.
So, this commit removes the documentation for this configure
option from `INSTALL.md`

Additionally, the work around for GASNet-EX bug 4494 is documented.