Resolve issue #565
This commit resolves issue #565: "`upcxx-run --help` fails in a build
directory" by special-casing `-h` and `--help` as arguments to
`upcxx-run`.

It is worth noting that the new logic suffers from the same deficiency
as described in issue 174 (arguments to the application are treated as
arguments to `upcxx-run`).  However, the same work around (use of `--`
at the end of the `upcxx-run` arguments) is effective.