Add an undocumented/optional `UPCXX_HAVE_CXX17` make flag
This optional flag asserts the compiler will accept a `-std=c++17`
argument (which need be not be in our default `CXXFLAGS`), and
that will activate sufficient C++17 feature support to build any tests
that currently require C++17 features (a deliberately unspecified
moving target).

This option is expected to be refined/expanded as additional tests
are added requiring features beyond C++11.