configure: enforce libstdc++ version for PGI/NVHPC
This commit adds a configure time probes for the GCC version specified
when an PGI or NVHPC compiler package was installed.  This provides
us with the means to reject installations with a too-old libstdc++.

The probe has been confirmed effective for a range of PGI compilers
from our current floor through the final release.  Similarly, the
first and current releases of NVHPC are correctly probed.  However,
to defend against the possible future renaming of the (probably
undocumented) identifier being probed, a failure to probe the
version will only generate a warning.

Resolves issue #348 RFE: libstdc++ enforcement for PGI compilers