Support AMD "AOCC" compilers on Linux
This commit updates `INSTALL.md` and `utils/system-checks.sh` to
document support for AMD's "AMD Optimizing C/C++ and Fortran
Compilers" on Linux x86_64 with a floor version corresponding to what
we currently test.  Logic in `utils/system-checks.sh` issues our
standard "has not been validated" or "is known to lack the support"
message if a too-old compiler is detected, depending on just how old
it is.

Resolves issue #586 Determine/document support for AMD compilers
(non-PrgEnv case)