Fix for in-build-tree utils w/o default network
Resolves issue 596

The problem in issue 596 was that when the utilities in `bin/` of
the build tree are generated by `configure`, they embed the value
of `$GASNET_CONDUIT` known to `configure`.  Unless there is a
system-specific default (`smp` on macOS and `aries` on an Cray XC)
this means that absence of `--with-default-network=...` results in
the utilities defaulting to an empty string for `UPCXX_NETWORK`.
This then fails validation against the list of all known networks.

There are two changes to the generation of `upcxx` and `upcxx-meta` to
address the problem.  The first is to allow and empty network value to
pass validation.  The second is to add to add login in the "may need
to build the required runtime" step to query the default network.
This is accomplished with the addition of a new (internal use) target
`echo-default-network` in `bld/Makefile`.

No changes are require to `upcxx-info`, which is only a wrapper for
the (fixed) `upcxx`.

No changes are required to `upcxx-run`, which does not suffer from
issue 596.