persona.hpp: simple code motion for LPC return path
This moves helper types used in the LPC return path out of persona
and into the detail namespace later in the file, in order to resolve
a dependency cycle in subsequent commits.

There are NO code changes in this commit, just code motion and
adjustment of the name used by the caller code.