Prevent escape sequence warnings in upcxx-run
This commit resolves issue #617 regarding warnings from Python 3.12
about unrecognized escape sequences in upcxx-run.

Starting with Python 3.12, the use of unrecognized backslash escape
sequences in string literals yield warnings, and future releases
are anticipated to produce errors.  Since even before Python 2.7,
documentation for `re` has documented use of "raw strings" as
recommended for regular expressions to prevent interpretation of
backslash sequences.  So, this commit simply puts that recommendation
in to practice for each regular expressions which contains the `\s`
and/or `\d` sequences.  Normal strings are notably still used in
regular expressions which include `\t` and `\n` (tab and newline,
respectively) where interpretation of the escape sequence is desired.