CCS tests: fix "hang" on failed test compilation
This commit adds the `-f` flag to an existing `rm -r` command on a
failure path to ensure that commands such as `make dev-check` will not
stop to wait for user input in the event certain test(s) fail to
compile.

Such a hang has been observed on at least one Linux system where the
linker was broken.