Makefile: repair `run` target.
This commit repairs a latent implicit assumption in the `run` target
in `bld/Makefile` that `#!/usr/bin/env python` in `util/upcxx-run`
will "just work".  That could fail if `python` is not in ones `$PATH`
or (less likely) if `/usr/bin/env` does not exist.

The fix is to explicitly invoke the Python interpreter via
`$(UPCXX_PYHON)`, similar to what the generated `bin/upcxx-run` in the
build directory does.