Remove support for aries-conduit / Cray XC
Following-up on the removal of aries-conduit in GASNet-EX 2025.8.0,
this commit removes support for aries-conduit and the Cray XC platform
from documentation and code.

The section of `docs/mpi-hybrid.md` which mentions use of `aprun` and
`srun` on the Cray XC has been repurposed for the HPE Cray EX, rather
than removed.